Base
Includes the main game file along with a battle, monster, and player classes. The user is prompted for a username then a random monster is selected and a battle starts. The user can attack or run from the monster. Battle will continue until either the player's HP reaches 0 or the monster's HP reaches 0. Running will prompt a new monster to be generated. Very basic game to get me back into programming and Python.